server: Add rocket.rs database server using hashmap
Implemented a basic Rust web server using rocket.rs that
stores User(id, username) in a HashMap managed by rocket
that is made available to the request handler functions
by passing it as a parameter.

A GET method is implemented for "/user/<id>" route to
get data about a user stored in the HashMap. If the requested
id does not exist, the error is caught using a catch(404)
catcher that returns a JSON with status and reason for the error.

A POST method is implemented at "/user" for storing user data
received as JSON in the HashMap if it does not exist already.

Signed-off-by: Andrei Cherechesu <andrei.cherechesu@gmail.com>